+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Changed
 - Debounce reconnects (#68).
+- Handle `INCOMING_TRACK_REMOVE` messages (#69).
 
 ### Fixed
 - Missed service request status messages (#70).