+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.1] - 2023-02-25
+### Fixed
+- `Accept-Language` header not set on mobile (#71).
+
 ## [1.0.0] - 2023-02-14
 ### Changed
 - Debounce reconnects (#68).